+++ include/common.h
@@ -6,6 +6,7 @@
 #define SUBDATANUM 2000000
 #define DATANUM (SUBDATANUM * MAX_THREADS)   /* total number of data */
 #define ACCESS(data) log(sqrt(data))
+#define ACCESSF(data) logf(sqrtf(data))
 
 enum RunningMode {
     LOCAL,