+++ src/original.cpp
@@ -2,7 +2,7 @@
 
 
 float sum(const float data[], const int len) {
-    double sum_value = 0;  // use double to guarantee the precision
+    double sum_value = 0.0;  // use double to guarantee the precision
     for (int i = 0; i < len; i++)
         sum_value += ACCESS(data[i]);
     return float(sum_value);